Refactor sitecore-jss package - break up layout-service into rest-layout-service and graphql-layout-service, similar to what was done for dictionary service - remove 'data-api' file to improve the dependency hierarchy (see Miro diagram) - break up util functions - isolate EE utils in 'experience-editor' file to prepare for Horizon work - break out logic for resolving app root into its own util for reuse in sitemap
