+++ packages/sitecore-jss/src/data-fetcher.ts
@@ -1,3 +1,9 @@
+import resolveUrl from './utils/resolve-url';
+
+/**
+ * Response data for an HTTP request sent to an API
+ * @template T the type of data model requested
+ */
 export interface HttpResponse<T> {
   /** HTTP status code of the response (i.e. 200, 404) */
   status: number;
@@ -8,12 +14,12 @@ export interface HttpResponse<T> {
 }
 
 /**
- * Interface for functions that fetch data asynchronously (i.e. from an API endpoint).
+ * Describes functions that fetch data asynchronously (i.e. from an API endpoint).
  * This interface conforms to Axios' public API, but is adaptable to other HTTP libraries and
  * fetch polyfills.
  * The interface implementation must:
  * - Support SSR
- * - Comply with the rules of REST by returning non-HTTP 200 status codes, not thrown exceptions (i.e. be a proper REST client)
+ * - Comply with the rules of REST by returning appropriate response status codes when there is an error instead of throwing exceptions.
  * - Send HTTP POST requests if `data` param is specified; GET is suggested but not required for data-less requests
  */
 export type HttpDataFetcher<T> = (
@@ -21,3 +27,44 @@ export type HttpDataFetcher<T> = (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   data?: { [key: string]: any }
 ) => Promise<HttpResponse<T>>;
+
+class ResponseError extends Error {
+  response: HttpResponse<unknown>;
+
+  constructor(message: string, response: HttpResponse<unknown>) {
+    super(message);
+
+    Object.setPrototypeOf(this, ResponseError.prototype);
+    this.response = response;
+  }
+}
+
+/**
+ * @param {HttpResponse<T>} response
+ */
+function checkStatus<T>(response: HttpResponse<T>) {
+  if (response.status >= 200 && response.status < 300) {
+    return response;
+  }
+
+  const error = new ResponseError(response.statusText, response);
+  throw error;
+}
+
+/**
+ * @param {string} url
+ * @param {HttpDataFetcher} fetcher
+ * @param {Object} params
+ */
+export function fetchData<T>(
+  url: string,
+  fetcher: HttpDataFetcher<T>,
+  params: { [key: string]: string | number | boolean } = {}
+) {
+  return fetcher(resolveUrl(url, params))
+    .then(checkStatus)
+    .then((response) => {
+      // axios auto-parses JSON responses, don't need to JSON.parse
+      return response.data;
+    });
+}
