+++ packages/sitecore-jss/src/experience-editor.ts
@@ -1,0 +1,17 @@
+import isServer from './utils/is-server';
+
+export const isExperienceEditorActive = (): boolean => {
+  if (isServer()) {
+    return false;
+  }
+  // eslint-disable-next-line
+  const sc = (window as any).Sitecore;
+  return Boolean(sc && sc.PageModes && sc.PageModes.ChromeManager);
+};
+
+export const resetExperienceEditorChromes = (): void => {
+  if (isExperienceEditorActive()) {
+    // eslint-disable-next-line
+    (window as any).Sitecore.PageModes.ChromeManager.resetChromes();
+  }
+};
