+++ packages/sitecore-jss/src/graphql-request-client.ts
@@ -49,6 +49,8 @@ export class GraphQLRequestClient {
         query,
         variables,
       });
+      console.log(query);
+      console.log(JSON.stringify(variables, null, 2));
       this.client
         .request(query, variables)
         .then((data: T) => {
