+++ packages/sitecore-jss/src/index.ts
@@ -1,27 +1,40 @@
-import * as dataApi from './data-api';
 import * as mediaApi from './media-api';
 import * as constants from './constants';
+import isServer from './utils/is-server';
+import resolveUrl from './utils/resolve-url';
 
-export { dataApi, mediaApi, constants };
+// TODO: these are deprecated and we should stop exporting them
+import { fetchRouteData, fetchPlaceholderData } from './layout/rest-layout-service';
+const dataApi = { fetchRouteData, fetchPlaceholderData };
+export { dataApi };
+
+export { mediaApi, constants, isServer, resolveUrl };
 
 export { default as debug, Debugger } from './debug';
 
-export { GraphQLRequestClient, GraphQLRequestClientConfig } from './graphql-request-client';
+export { isExperienceEditorActive, resetExperienceEditorChromes } from './experience-editor';
 
-export { LayoutServiceRequestOptions } from './data-api';
+// generic API access abstractions
+export { HttpDataFetcher, HttpResponse } from './data-fetcher';
+
+// API access implementations
+export { GraphQLRequestClient, GraphQLRequestClientConfig } from './graphql-request-client';
 
 export { AxiosDataFetcher, AxiosDataFetcherConfig } from './axios-fetcher';
 
-export { LayoutService } from './layout/layout-service';
+export { getAppRoot } from './utils/app-root-resolver';
+
+// i18n
+export { DictionaryPhrases, DictionaryService } from './i18n/dictionary-service';
+
 export {
-  RestLayoutService,
-  RestLayoutServiceConfig,
-  DataFetcherResolver,
-} from './layout/rest-layout-service';
-export { GraphQLLayoutService, GraphQLLayoutServiceConfig } from './layout/graphql-layout-service';
+  GraphQLDictionaryService,
+  GraphQLDictionaryServiceConfig,
+} from './i18n/graphql-dictionary-service';
 
-export { isExperienceEditorActive, isServer, resetExperienceEditorChromes, urlUtil } from './util';
+export { RestDictionaryService, RestDictionaryServiceConfig } from './i18n/rest-dictionary-service';
 
+// layout
 export {
   LayoutServiceData,
   LayoutServicePageState,
@@ -40,12 +53,13 @@ export {
 
 export { getFieldValue, getChildPlaceholder } from './layout/utils';
 
-export { HttpDataFetcher, HttpResponse } from './data-fetcher';
+export { LayoutService } from './layout/layout-service';
 
-// i18n
-export { DictionaryPhrases, DictionaryService } from './i18n/dictionary-service';
 export {
-  GraphQLDictionaryService,
-  GraphQLDictionaryServiceConfig,
-} from './i18n/graphql-dictionary-service';
-export { RestDictionaryService, RestDictionaryServiceConfig } from './i18n/rest-dictionary-service';
+  RestLayoutService,
+  RestLayoutServiceConfig,
+  DataFetcherResolver,
+  LayoutServiceRequestOptions,
+} from './layout/rest-layout-service';
+
+export { GraphQLLayoutService, GraphQLLayoutServiceConfig } from './layout/graphql-layout-service';
